# dto/admin.go
package dto

import (
	"github.com/gin-gonic/gin"
	"github.com/lhc328/go_gateway_demo/public"
	"time"
)

type AdminInfoOutput struct {
	ID int `json:"id"`
	Name string `json:"name"'`
	LoginTime time.Time `json:login_time`
	Avatar string `json:"avatar"`
	Introduction string `json:introduction`
	Roles []string `json:"roles"`
}

type ChangePassInput struct {
	Password string `json:"password" form:"password" comment:"密码" example:"123456" validate:"required"` //密码
}

func (param *ChangePassInput) BindValidParam(c *gin.Context) error {
	return public.DefaultGetValidParams(c, param)
}